requirements.txt
alembic==1.4.2
astroid==2.5.3
autopep8==1.5.6
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
wheel==0.36.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.6.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.7.4
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
